# backend-express/swagger.yaml
openapi: 3.0.0
info:
  title: Kopi Haus Dashboard API
  version: 1.0.0
  description: API documentation for the Kopi Haus Dashboard backend application.
servers:
  - url: http://localhost:5001/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Server error
    Outlet:
      type: object
      required:
        - name
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c1
        name:
          type: string
          example: Main Branch
        address:
          type: string
          example: 123 Coffee St, City
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
...
    Expense:
      type: object
      required:
        - outlet_id
        - category
        - amount
        - noted_by
        - spent_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c27
        outlet_id:
          type: string
          description: ID of the outlet where the expense was incurred
          example: 60d5ec49f8c7a40015a7b8c1
        shift_id:
          type: string
          nullable: true
          description: ID of the shift during which the expense was incurred (if any)
          example: 60d5ec49f8c7a40015a7b8c12
        category:
          type: string
          description: Category of the expense (e.g., Utilities, Rent, Supplies)
          example: Supplies
        amount:
          type: number
          format: float
          description: Amount of the expense
          example: 50000
        noted_by:
          type: string
          description: ID of the user who noted the expense
          example: 60d5ec49f8c7a40015a7b8c13
        spent_at:
          type: string
          format: date-time
          description: Timestamp when the expense was incurred
          example: '2023-10-26T15:00:00.000Z'
        note:
          type: string
          nullable: true
          description: Additional notes about the expense
          example: Purchase of cleaning supplies
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T15:00:00.000Z'
paths:
...
  /outlets:
    get:
      summary: Get all outlets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of outlets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Outlet'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new outlet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: New Outlet Name
                address:
                  type: string
                  example: New Outlet Address
      responses:
        '201':
          description: Outlet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /outlets/{id}:
    get:
      summary: Get an outlet by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the outlet to retrieve
      responses:
        '200':
          description: Outlet data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '404':
          description: Outlet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update an outlet by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the outlet to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Updated Outlet Name
                address:
                  type: string
                  example: Updated Outlet Address
      responses:
        '200':
          description: Outlet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Outlet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an outlet by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the outlet to delete
      responses:
        '200':
          description: Outlet deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Outlet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /expenses:
    get:
      summary: Get all expenses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '500':
          description: Server error

    AuditLog:
      type: object
      required:
        - action
        - entity
        - created_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c28
        actor_id:
          type: string
          nullable: true
          example: 60d5ec49f8c7a40015a7b8c13
          description: ID of the user who performed the action, null for system actions
        action:
          type: string
          description: Type of action performed (e.g., 'create', 'update', 'delete', 'login')
          example: create
        entity:
          type: string
          description: Name of the entity affected (e.g., 'User', 'Product', 'Sale')
          example: User
        entity_id:
          type: string
          nullable: true
          description: ID of the entity affected
          example: 60d5ec49f8c7a40015a7b8c13
        meta_json:
          type: object
          description: Additional metadata in JSON format
          example: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T16:00:00.000Z'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - outlet_id
                - category
                - amount
                - noted_by
              properties:
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                shift_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c12
                category:
                  type: string
                  example: Utilities
                amount:
                  type: number
                  format: float
                  example: 150000
                noted_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c13
                note:
                  type: string
                  nullable: true
                  example: Electricity bill
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Invalid input, outlet ID, shift ID, or user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expenses/{id}:
    get:
      summary: Get an expense by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the expense to retrieve
      responses:
        '200':
          description: Expense data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update an expense by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the expense to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                shift_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c12
                category:
                  type: string
                  example: Rent
                amount:
                  type: number
                  format: float
                  example: 2000000
                noted_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c13
                note:
                  type: string
                  nullable: true
                  example: Monthly rent payment
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Invalid input, outlet ID, shift ID, or user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an expense by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the expense to delete
      responses:
        '200':
          description: Expense deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...